Move reload command into LocalExtension for local-dev use commands only, that are optionally loaded if running in local development mode
